Add TS file import script
The TS file is Qt Trasnlation file, it actually is XML file.
bin/ts2sql.py parse the XML format and iterate through the message
field to write the value to the PostgreSQL database.

 - bin/ts2sql.py zh_CN.ts

And this script also will check the language type of TS file if
detected its basename include language code.

Signed-off-by: Burgess Chang <bsc@brsvh.org>